# python/openapi.yaml
openapi: 3.0.1
info:
  description: KnoxPothole API Documentation
  version: "1.0.0"
  title: KnoxPothole API
servers:
  - url: /api/v1
paths:
  /reports/get:
    get:
      operationId: reports.getAllReports
      tags:
        - Get
        - App
      summary: Reports Get
      description: Reports Get
      responses:
        '200':
          description: Successfully received list of reports
          content:
            application/json:
              schema:
                title: Array of Reports
                type: array
                items:
                  properties:
                    location:
                      type: object
                      description: The location of an object
                    status_id:
                      type: integer
                      description: The status of pothole filling
                    zone:
                      type: string
                      description: Zone of pothole location
                    priority:
                      type: string
                      description: The priority
                    reporter:
                      type: string
                      description: The reporter's name
  /reports/post:
    post:
      operationId: reports.postReport
      tags:
        - Post
        - App
      summary: Add new report
      description: Add new report
      responses:
        '201':
          description: Successfully added a report
      requestBody:
        description: Add new report
        required: true
        content:
          application/json:
            schema:
              required:
                - location
                - status_id
                - zone
              type: object
              properties:
                location:
                  type: object
                  description: The location of an object
                status_id:
                  type: integer
                  description: The status of pothole filling
                zone:
                  type: string
                  description: Zone of pothole location
                priority:
                  type: string
                  description: The priority
                reporter:
                  type: string
                  description: The reporter's name
  /location/post/gps:
    post:
      operationId: reports.postLocationGPS
      tags:
        - Post
        - App
      summary: Add new location by GPS coordinates
      description: Add new location by GPS coordinates
      responses:
        '201':
          description: Succesfully added location
      requestBody:
        description: Add new location
        required: true
        content:
          application/json:
            schema:
              required:
                - latitude
                - longitude
              type: object
              properties:
                latitude:
                  type: string
                  description: Latitude
                longitude:
                  type: string
                  description: Longitude
  /location/post/address:
      post:
        operationId: reports.postLocationAddress
        tags:
          - Post
          - App
        summary: Add new location by address
        description: Add new location by address
        responses:
          '201':
            description: Succesfully added location
        requestBody:
          description: Add new location
          required: true
          content:
            application/json:
              schema:
                required:
                  - address
                type: object
                properties:
                  address:
                    type: string
                    description: Address
  /workorders/get:
    get:
      operationId: reports.getAllWorkorders
      tags:
        - Get
        - App
      summary: Get all work orders
      description: Get all work orders
      responses:
        '200':
          description: Successfully received all work orders
          content:
            application/json:
              schema:
                title: Array of work orders
                type: array
                items:
                  properties:
                    wo:
                      type: string
                      description: W/O ID for documentation
                    request_id:
                      type: object
                      description: Request associated with the workorder
  /location/get/gps:
    get:
      operationId: reports.getLocationGPS
      tags:
        - Get
        - App
      summary: Get the address at the GPS coordinates
      description: Get the address at the GPS coordinates
      parameters:
        - name: longitude
          in: query
          required: true
          description: Longitude
          schema:
            type: string
        - name: latitude
          in: query
          required: true
          description: Latitude
          schema:
            type: string

      responses:
        '200':
          description: Successfully retrieved the address
          content:
            application/json:
              schema:
                title: Array of work orders
                type: object
                items:
                  properties:
                    address:
                      type: string
                      description: Address of the location
  /reports/get/status:
    get:
      operationId: reports.getReportStatus
      tags:
        - Get
        - App
      summary: Get reports of the status
      description: Get reports with the inputted status
      responses:
        '200':
          description: Successfully received list of reports
          content:
            application/json:
              schema:
                title: Array of Reports
                type: array
                items:
                  properties:
                    location:
                      type: object
                      description: The location of an object
                    status_id:
                      type: integer
                      description: The status of pothole filling
                    zone:
                      type: string
                      description: Zone of pothole location
                    priority:
                      type: string
                      description: The priority
                    reporter:
                      type: string
                      description: The reporter's name
  /workorders/get/status:
    get:
      operationId: reports.getWorkorderStatus
      tags:
        - Get
        - App
      summary: Get all work orders with the given status
      description: Get all work orders with the given status
      responses:
        '200':
          description: Successfully received all work orders
          content:
            application/json:
              schema:
                title: Array of work orders
                type: array
                items:
                  properties:
                    wo:
                      type: string
                      description: W/O ID for documentation
                    request_id:
                      type: object
                      description: Request associated with the workorder
  /reports/get/location:
    get:
      operationId: reports.getReportLocation
      tags:
        - Get
        - App
      summary: Get reports of the location
      description: Get reports with the given address
      responses:
        '200':
          description: Successfully received list of reports
          content:
            application/json:
              schema:
                title: Array of Reports
                type: array
                items:
                  properties:
                    location:
                      type: object
                      description: The location of an object
                    status_id:
                      type: integer
                      description: The status of pothole filling
                    zone:
                      type: string
                      description: Zone of pothole location
                    priority:
                      type: string
                      description: The priority
                    reporter:
                      type: string
                      description: The reporter's name
  /workorders/get/location:
    get:
      operationId: reports.getWorkorderLocation
      tags:
        - Get
        - App
      summary: Get work orders of the location
      description: Get work orders with the given address
      responses:
        '200':
          description: Successfully received list of work orders
          content:
            application/json:
              schema:
                title: Array of Work Orders
                type: array
                items:
                  properties:
                    wo:
                      type: object
                      description: Work Order
                    
  /jobs/getoptimal:
    get:
      operationId: reports.getOptimalJobs()
      tags:
        - Get
        - App
      summary: Get list of jobs in the optimal path. Max 23 points.
      description: Get list of jobs in the optimal path. Max 23 points.
      responses:
        '200':
          description: Successfully got list of jobs
          content:
            application/json:
              schema:
                title: Array of jobs
                type: array
                items:
                  properties:
                    workorder:
                      type: object
                      description: Workorder associated with the job